Add wallet create and update

+++ src/main/kotlin/trackerr/rezyfr/dev/data/table/WalletTable.kt
@@ -6,7 +6,7 @@ object WalletTable : Table() {
 
     val id = integer("id").autoIncrement()
     val name = varchar("name", 128)
-    val balance = integer("balance")
+    val balance = long("balance")
     val userEmail = varchar("user_email", 128).references( UserTable.email)
     val color = long("color")
     val icon = varchar("icon", 128)
